Minor changes in netjack2
git-svn-id: http://subversion.jackaudio.org/jack/jack2/trunk/jackmp@2751 0c269be4-1314-0410-8aa9-9f06e86f4224

+++ common/JackNetUnixSocket.h
@@ -80,7 +80,7 @@ namespace Jack
         int GetOption ( int level, int optname, void* optval, socklen_t* optlen );
 
         //timeout
-        int SetTimeOut ( int& msec );
+        int SetTimeOut ( float& msec );
 
         //local loop
         int SetLocalLoop();
